+++ examples/quick_start.sh
@@ -59,6 +59,7 @@ docker run --rm -it \
     "${avatars_uri[@]}" \
     "${logo_uri[@]}" \
     "${background_image_uri[@]}" \
+    "${default_user_image_uri[@]}" \
     "${env_vars[@]}" \
     "${args[@]}" \
     cartoonman/envisaged-redux:latest